While statement added to consumer to wait for messages. Data type of one column changed, it was bad. JAR added to Dockerfile to load data correctly from main.py

+++ postgres/init.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS bus_traffic_processed (
     VehicleCoordinates DOUBLE PRECISION,
     NextStopPointName DOUBLE PRECISION,
     ArrivalProximityText DOUBLE PRECISION,
-    DistanceFromStop INTEGER,
+    DistanceFromStop DOUBLE PRECISION,
     Delay DOUBLE PRECISION,
     PRIMARY KEY (RecordedAtTime_year, RecordedAtTime_month, RecordedAtTime_day, RecordedAtTime_hour, RecordedAtTime_minute, RecordedAtTime_second)
 );
